Made unit tests compile again.
